requirements.txt
psycopg2-binary
requests
fastapi
uvicorn
python-dotenv
asyncpg
tqdm
httpx
pytest
